+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Disable Chunked Mode for sending data via HttpURLConnection
+- Non-finite numeric values are serialized as JSON null in reported events, and a special field is added for supportability.
 
 ## 3.0.0 [Release date: 2022-12-06]
 [GitHub Releases](https://github.com/Dynatrace/openkit-java/releases/tag/v3.0.0)